clean up storybook
